requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.10.0,<3.0.0
openai-agents>=0.2.0
google-genai
qdrant-client
google-generativeai